Move URL logic to its own file
